contractor-system/Backend/requirements.txt
--only-binary pydantic-core

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9

# Supabase
supabase==2.7.4

# Auth & Settings
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Validation
pydantic==2.9.2          # ↑ במקום 2.6.1
pydantic-settings==2.2.0
email-validator==2.1.0

# Email Service
sendgrid==6.11.0

# File Handling
python-magic==0.4.27
pillow==11.0.0

# PDF Generation
reportlab==4.2.5

# Utilities
httpx==0.26.0
